Add second test wiki database
To enable testing multi-project cohorts, we add a new test wiki database
called "wiki2_testing".  This is created and grants all access to the
wikimetrics user.

Furthermore, this change would be incorrect if ::wikimetrics::debug is
false, so I added a check for that around all testing database setup.

Adding new define wikimetrics::database::create to abstract out
creation of databases and granting access to them.  This DRYs the
wikimetrics::database a bit.

Change-Id: I32e1520be9fb507a95b829eaf38a7018257d5c66