Make hourly backup keep around known-good full backups in case of issues
Since we cannot fully control when redis or wikimetrics modify the
files that we want to backup, tar might act on files that are changed
as they get read. This lead to broken/partial backups.

To account for this limitation, we change the backup strategy to work
around it upon need:

If the hourly backup does not run into problems, it produces a single
tar-ball containing new, full backup. Just as it did before.

However, if the hourly backup does run into issues (like the above
read/write issue), the hourly script leaves the old, known-good, full
backup in place, and additionally provides the new, more
up-to-date (but maybe not full) backup tar-ball. In this case, the
daily backup script copies over both tarballs, so we can recover as
much data as possible.

Bug: 68731
Change-Id: I68845721a902ae5208a677c072355289fbd0c835